+++ repo/src/main/java/com/parashift/onlyoffice/scripts/DownloadAs.java
@@ -1,8 +1,6 @@
 package com.parashift.onlyoffice.scripts;
 
-import com.parashift.onlyoffice.util.ConvertManager;
-import com.parashift.onlyoffice.util.UrlManager;
-import com.parashift.onlyoffice.util.Util;
+import com.parashift.onlyoffice.util.*;
 import org.alfresco.model.ContentModel;
 import org.alfresco.repo.download.DownloadModel;
 import org.alfresco.repo.download.DownloadStatusUpdateService;
@@ -18,18 +16,16 @@ import org.alfresco.util.TempFileProvider;
 import org.apache.commons.compress.archivers.zip.ZipArchiveEntry;
 import org.apache.commons.compress.archivers.zip.ZipArchiveOutputStream;
 import org.apache.commons.io.IOUtils;
+import org.apache.http.HttpEntity;
+import org.apache.http.util.EntityUtils;
 import org.json.JSONArray;
 import org.json.JSONObject;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.extensions.surf.util.URLEncoder;
 import org.springframework.extensions.webscripts.*;
 import org.springframework.stereotype.Component;
 
 import java.io.*;
-import java.net.HttpURLConnection;
-import java.net.URL;
 import java.text.MessageFormat;
 import java.util.*;
 import java.util.zip.Deflater;
@@ -45,8 +41,6 @@ public class DownloadAs extends AbstractWebScript {
     private static final String TEMP_FILE_PREFIX = "alf";
     private static final String ZIP_EXTENSION = ".zip";
     private final static String CONTENT_DOWNLOAD_API_URL = "slingshot/node/content/{0}/{1}/{2}/{3}";
-
-    private Logger logger = LoggerFactory.getLogger(this.getClass());
 
     @Autowired
     ConvertManager converterService;
@@ -78,6 +72,9 @@ public class DownloadAs extends AbstractWebScript {
     @Autowired
     UrlManager urlManager;
 
+    @Autowired
+    RequestManager requestManager;
+
     @Override
     public void execute(WebScriptRequest request, WebScriptResponse response) throws IOException {
         try {
@@ -89,7 +86,7 @@ public class DownloadAs extends AbstractWebScript {
                 JSONObject data = requestDataJson.getJSONObject(0);
 
                 NodeRef node = new NodeRef(data.getString("nodeRef"));
-                String outputType = data.getString("outputType");
+                final String outputType = data.getString("outputType");
 
                 if (permissionService.hasPermission(node, PermissionService.READ) != AccessStatus.ALLOWED) {
                     throw new AccessDeniedException("Access denied. You do not have the appropriate permissions to perform this operation. NodeRef= " + node.toString());
@@ -103,13 +100,16 @@ public class DownloadAs extends AbstractWebScript {
                     contentURL = getDownloadAPIUrl(node, docTitle);
                 } else {
                     String downloadUrl = converterService.convert(util.getKey(node), currentExt, outputType, urlManager.getContentUrl(node), region);
-                    String docTitle = util.getTitleWithoutExtension(node) + "." + outputType;
+                    final String docTitle = util.getTitleWithoutExtension(node) + "." + outputType;
 
-                    URL url = new URL(urlManager.replaceDocEditorURLToInternal(downloadUrl));
-                    HttpURLConnection connection = (HttpURLConnection) url.openConnection();
-                    try (InputStream inputStream = connection.getInputStream()) {
-                        contentURL = createDownloadNode(docTitle, mimetypeService.getMimetype(outputType), inputStream, connection.getContentLength(), 1);
-                    }
+                    contentURL = requestManager.executeRequestToDocumentServer(downloadUrl, new RequestManager.Callback<String>() {
+                        public String doWork(HttpEntity httpEntity) throws IOException {
+                            byte[] bytes = EntityUtils.toByteArray(httpEntity);
+                            InputStream inputStream = new ByteArrayInputStream(bytes);
+
+                            return createDownloadNode(docTitle, mimetypeService.getMimetype(outputType), inputStream, bytes.length, 1);
+                        }
+                    });
                 }
             } else {
                 File zip = TempFileProvider.createTempFile(TEMP_FILE_PREFIX, ZIP_EXTENSION);
@@ -145,12 +145,13 @@ public class DownloadAs extends AbstractWebScript {
                             String downloadUrl = converterService.convert(util.getKey(node), currentExt, outputType, urlManager.getContentUrl(node), region);
                             String docTitle = util.getTitleWithoutExtension(node) + "." + outputType;
 
-                            URL url = new URL(urlManager.replaceDocEditorURLToInternal(downloadUrl));
-                            HttpURLConnection connection = (HttpURLConnection) url.openConnection();
-                            try (InputStream inputStream = connection.getInputStream()) {
-                                out.putArchiveEntry(new ZipArchiveEntry(docTitle));
-                                totalSize = totalSize + IOUtils.copyLarge(inputStream, out);
-                            }
+                            out.putArchiveEntry(new ZipArchiveEntry(docTitle));
+
+                            totalSize += requestManager.executeRequestToDocumentServer(downloadUrl, new RequestManager.Callback<Long>() {
+                                public Long doWork(HttpEntity httpEntity) throws IOException {
+                                    return IOUtils.copyLarge(httpEntity.getContent(), out);
+                                }
+                            });
                         }
 
                         out.closeArchiveEntry();