+++ repo/src/main/java/com/parashift/onlyoffice/util/JwtManager.java
@@ -3,6 +3,7 @@ package com.parashift.onlyoffice.util;
 import com.auth0.jwt.JWT;
 import com.auth0.jwt.algorithms.Algorithm;
 import com.auth0.jwt.interfaces.DecodedJWT;
+import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.json.JSONObject;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -30,19 +31,19 @@ public class JwtManager {
         return configManager.demoActive() || configManager.get("jwtsecret") != null && !((String)configManager.get("jwtsecret")).isEmpty();
     }
 
-    public String createToken(JSONObject payload) throws Exception {
+    public String createToken(JSONObject payload) throws JsonProcessingException {
         Map<String, ?> payloadMap = objectMapper.readValue(payload.toString(), Map.class);
 
         return createToken(payloadMap);
     }
 
-    public String createToken(Object payload) throws Exception {
+    public String createToken(Object payload) {
         Map<String, ?> payloadMap = objectMapper.convertValue(payload, Map.class);
 
         return createToken(payloadMap);
     }
 
-    public String createToken(Map<String, ?> payloadMap) throws Exception {
+    public String createToken(Map<String, ?> payloadMap) {
         Algorithm algorithm = Algorithm.HMAC256(getJwtSecret());
 
         String token = JWT.create()